refactor: split deals page into sub-components (2887 -> 1684 lines)
Extract helpers, presentational components, KPI row, and filter bar from
the monolithic deals index.jsx into dedicated files under components/.
All 1059 tests pass; build succeeds; no business logic changed.

- dealHelpers.js: pure utility functions (formatting, date keys, display helpers)
- DealPresentational.jsx: StatusPill, LoanerBadge, Pill, CustomerDisplay,
  ServiceLocationTag, DraftReminderBanner, ErrorAlert, DealCoreSnapshot,
  SheetSummaryRow, SheetViewTable, DeleteConfirmModal
- DealsKpiRow.jsx: KPI summary cards
- DealsFilterBar.jsx: status tabs, search, month filter, advanced filters, view toggle
