+++ src/pages/deals/index.jsx
@@ -8,9 +8,8 @@ import ExportButton from '../../components/common/ExportButton'
 import NewDealModal from './NewDealModal'
 import EditDealModal from './components/EditDealModal'
 import DealDetailDrawer from './components/DealDetailDrawer'
-import { money0, pct1, titleCase, prettyPhone } from '../../lib/format'
+import { titleCase } from '../../lib/format'
 import ScheduleBlock from '../../components/deals/ScheduleBlock'
-import { formatEtMonthDay, toSafeDateForTimeZone } from '../../utils/scheduleDisplay'
 import { getReopenTargetStatus } from '@/utils/jobStatusTimeRules.js'
 import { calculateDealKPIs, getDealFinancials } from '../../utils/dealKpis'
 
@@ -21,757 +20,43 @@ import Button from '../../components/ui/Button'
 import Icon from '../../components/ui/Icon'
 import { useToast } from '../../components/ui/ToastProvider'
 
+// Extracted sub-components
+import {
+  StatusPill,
+  LoanerBadge,
+  Pill,
+  CustomerDisplay,
+  ServiceLocationTag,
+  DraftReminderBanner,
+  ErrorAlert,
+  DealCoreSnapshot,
+  SheetSummaryRow,
+  SheetViewTable,
+  DeleteConfirmModal,
+} from './components/DealPresentational'
+import DealsKpiRow from './components/DealsKpiRow'
+import DealsFilterBar from './components/DealsFilterBar'
+
+// Extracted helpers
+import {
+  safeJsonStringify,
+  formatCreatedShort,
+  formatStaffName,
+  toFiniteNumberOrNull,
+  formatMoney0OrDash,
+  getEtDayKey,
+  getDealPromiseIso,
+  getPromiseDayKey,
+  getDealProductLabelSummary,
+  getDealVehicleDisplay,
+  getDealPrimaryRef,
+  getDisplayPhone,
+  isDealsDebugEnabled,
+} from './components/dealHelpers'
+
 // Feature flag for Simple Agenda
 const SIMPLE_AGENDA_ENABLED =
   String(import.meta.env.VITE_SIMPLE_CALENDAR || '').toLowerCase() === 'true'
-
-const isDealsDebugEnabled = () =>
-  import.meta.env.DEV &&
-  (import.meta.env.VITE_DEBUG_DEALS_LIST === 'true' ||
-    (typeof window !== 'undefined' && window.localStorage?.getItem('debug:deals') === '1'))
-
-const safeJsonStringify = (value) => {
-  try {
-    return JSON.stringify(value)
-  } catch {
-    try {
-      return String(value)
-    } catch {
-      return '[unserializable]'
-    }
-  }
-}
-
-const formatCreatedShort = (input) => {
-  if (!input) return null
-  try {
-    const d = input instanceof Date ? input : new Date(input)
-    if (Number.isNaN(d.getTime())) return null
-    return d.toLocaleDateString('en-US', { month: 'short', day: 'numeric', year: 'numeric' })
-  } catch {
-    return null
-  }
-}
-
-// ✅ UPDATED: StatusPill with enhanced styling
-const StatusPill = ({ status }) => {
-  const statusColors = {
-    draft: 'bg-accent/50 text-foreground',
-    pending: 'bg-blue-500/10 text-blue-200',
-    in_progress: 'bg-amber-500/10 text-amber-200',
-    completed: 'bg-emerald-500/10 text-emerald-200',
-    cancelled: 'bg-red-500/10 text-red-200',
-  }
-  const color = statusColors?.[status] || 'bg-accent/50 text-foreground'
-  const displayStatus = status?.replace('_', ' ')?.toUpperCase() || 'UNKNOWN'
-
-  return <span className={`px-2 py-1 rounded text-xs font-medium ${color}`}>{displayStatus}</span>
-}
-
-// Small badge for loaner status in lists
-const LoanerBadge = ({ deal }) => {
-  const dueShort = deal?.loaner_eta_return_date
-    ? formatEtMonthDay(deal.loaner_eta_return_date)
-    : null
-  return (
-    <span className="inline-flex items-center px-2 py-0.5 rounded text-xs font-medium bg-accent/50 text-foreground border border-border">
-      <Icon name="Car" size={12} className="mr-1" />
-      {deal?.loaner_number ? `#${deal.loaner_number}` : 'Loaner'}
-      {dueShort ? ` • Due ${dueShort}` : ''}
-    </span>
-  )
-}
-
-// Helper: Get display phone from deal, preferring normalized E.164 field
-const getDisplayPhone = (deal) => {
-  // Prefer customer_phone_e164 (normalized), fallback to customer_phone, then customer_mobile
-  const phone = deal?.customer_phone_e164 || deal?.customer_phone || deal?.customer_mobile || ''
-  return prettyPhone(phone) || '—'
-}
-
-// ✅ ADDED: Helper to format names as "Lastname, F."
-const formatStaffName = (fullName) => {
-  if (!fullName) return ''
-  const parts = fullName?.trim()?.split(/\s+/)
-  if (parts?.length === 0) return ''
-  if (parts?.length === 1) return parts[0]
-
-  const lastName = parts[parts.length - 1]
-  const firstInitial = parts[0]?.[0] ?? ''
-
-  return `${lastName}, ${firstInitial}.`
-}
-
-const getDealPromiseIso = (deal) => {
-  const explicit = deal?.next_promised_iso
-  if (explicit) return explicit
-
-  try {
-    if (Array.isArray(deal?.job_parts)) {
-      const dates = deal.job_parts
-        .map((p) => p?.promised_date)
-        .filter(Boolean)
-        .sort()
-      return dates?.[0] || null
-    }
-  } catch {
-    // ignore
-  }
-
-  return null
-}
-
-const getEtDayKey = (isoOrDate) => {
-  const d = toSafeDateForTimeZone(isoOrDate)
-  if (!d) return ''
-  return new Intl.DateTimeFormat('en-CA', {
-    timeZone: 'America/New_York',
-    year: 'numeric',
-    month: '2-digit',
-    day: '2-digit',
-  }).format(d)
-}
-
-const getPromiseDayKey = (deal) => {
-  const iso = getDealPromiseIso(deal)
-  if (!iso) return ''
-  // Deal service may provide YYYY-MM-DD or an ISO string; normalize to day key.
-  return String(iso).slice(0, 10)
-}
-
-const normalizeProductName = (name) => {
-  if (!name) return ''
-  return String(name)
-    .replace(/\s*\([^)]*\)\s*/g, ' ')
-    .replace(/\s+/g, ' ')
-    .trim()
-}
-
-const abbreviateProductName = (name) => {
-  const raw = normalizeProductName(name)
-  if (!raw) return ''
-
-  const lower = raw.toLowerCase()
-  if (lower.includes('paint protection film')) return 'PPF'
-  if (lower.includes('tint')) return 'Tint'
-  if (lower.includes('ceramic')) return 'Ceramic'
-  if (lower.includes('rust')) return 'Rust'
-  if (/(^|\b)ppf(\b|$)/i.test(raw)) return 'PPF'
-
-  const words = raw.split(' ').filter(Boolean)
-  if (words.length >= 2 && words[0].length <= 12) return titleCase(words[0])
-
-  const maxLen = 14
-  if (raw.length <= maxLen) return titleCase(raw)
-  return `${titleCase(raw.slice(0, maxLen - 1))}…`
-}
-
-const getDealProductLabelSummary = (deal, maxLabels = 3) => {
-  const parts = Array.isArray(deal?.job_parts) ? deal.job_parts : []
-
-  const byKey = new Map()
-  for (const part of parts) {
-    const opCodeRaw = part?.product?.op_code || part?.product?.opCode || part?.op_code || ''
-    const opCode = String(opCodeRaw || '')
-      .trim()
-      .toUpperCase()
-    const name =
-      part?.product?.name ||
-      part?.product_name ||
-      part?.productLabel ||
-      part?.product?.label ||
-      part?.product_id ||
-      ''
-
-    const abbr = opCode || abbreviateProductName(name)
-    if (!abbr) continue
-
-    const keyRaw = part?.product?.id || part?.product_id || abbr
-    const key = String(keyRaw).toLowerCase()
-
-    const qtyRaw = part?.quantity_used ?? part?.quantity ?? 1
-    const qtyNum = Number(qtyRaw)
-    const qty = Number.isFinite(qtyNum) && qtyNum > 0 ? qtyNum : 1
-
-    const isOffSite = !!(part?.is_off_site || part?.service_type === 'vendor')
-
-    const existing = byKey.get(key)
-    if (!existing) {
-      byKey.set(key, { label: abbr, qty, isOffSite })
-    } else {
-      existing.qty += qty
-      existing.isOffSite = existing.isOffSite || isOffSite
-    }
-  }
-
-  const items = Array.from(byKey.values()).sort((a, b) => {
-    if (a.isOffSite !== b.isOffSite) return a.isOffSite ? -1 : 1
-    return a.label.localeCompare(b.label)
-  })
-
-  const formatted = items.map((it) => (it.qty > 1 ? `${it.label}×${it.qty}` : it.label))
-  const clipped = formatted.slice(0, maxLabels)
-  const extraCount = Math.max(0, formatted.length - clipped.length)
-  return { labels: clipped, extraCount }
-}
-
-const getDealVehicleDisplay = (deal) => {
-  const year = deal?.vehicle?.year
-  const make = deal?.vehicle?.make
-  const model = deal?.vehicle?.model
-  const stock =
-    deal?.vehicle?.stock_number ||
-    deal?.vehicle?.vin ||
-    deal?.stock_no ||
-    deal?.stockNumber ||
-    deal?.vin ||
-    null
-
-  const ymm = [year, make, model].filter(Boolean).join(' ').trim()
-  const vehicleLabel = ymm ? titleCase(ymm) : ''
-
-  const rawDesc = (deal?.vehicle_description || '').trim()
-  const title = ((deal?.title || deal?.description || '') + '').trim()
-  const jobNumber = ((deal?.job_number || deal?.jobNumber || '') + '').trim()
-  const descLower = rawDesc.toLowerCase()
-
-  const isDuplicativeDesc =
-    !rawDesc ||
-    (title && descLower === title.toLowerCase()) ||
-    (jobNumber &&
-      (descLower === jobNumber.toLowerCase() || descLower.includes(jobNumber.toLowerCase())))
-
-  // Vehicle-first:
-  // 1) Prefer structured year/make/model
-  // 2) Then accept vehicle_description only if it doesn't duplicate job/title
-  // 3) Then fall back to stock/vin snippet
-  if (vehicleLabel) {
-    return {
-      main: vehicleLabel,
-      stock: stock || null,
-      title: `${vehicleLabel}${stock ? ` • Stock: ${stock}` : ''}`,
-      isMissing: false,
-    }
-  }
-
-  if (!isDuplicativeDesc) {
-    const descLabel = titleCase(rawDesc)
-    return {
-      main: descLabel,
-      stock: null,
-      title: descLabel,
-      isMissing: false,
-    }
-  }
-
-  if (stock) {
-    const stockLabel = `Stock: ${stock}`
-    return {
-      main: stockLabel,
-      stock: null,
-      title: stockLabel,
-      isMissing: false,
-    }
-  }
-
-  return { main: '—', stock: null, title: '—', isMissing: true }
-}
-
-const Pill = ({ children, className = '' }) => (
-  <span
-    className={`inline-flex items-center rounded-full px-2.5 py-1 text-xs font-medium bg-accent/50 text-foreground ${className}`}
-  >
-    {children}
-  </span>
-)
-
-// ✅ ADDED: Customer display helper so table renders without missing component
-const CustomerDisplay = ({ deal }) => {
-  if (!deal) return <span className="text-sm text-muted-foreground">—</span>
-
-  const rawName =
-    deal?.customer_name ||
-    deal?.customerName ||
-    deal?.vehicle?.owner_name ||
-    deal?.customer_email ||
-    deal?.customerEmail ||
-    '—'
-  const name = rawName // already title-cased in DB where applicable
-  const email = deal?.customer_email || deal?.customerEmail || deal?.vehicle?.owner_email || ''
-  const tags = Array.isArray(deal?.work_tags) ? deal.work_tags : []
-  const title = [name, email, tags.length ? `Tags: ${tags.join(', ')}` : null]
-    .filter(Boolean)
-    .join(' • ')
-
-  return (
-    <div className="flex flex-col gap-1" title={title}>
-      <span
-        className="text-sm font-medium text-foreground"
-        data-testid={deal?.id ? `deal-customer-name-${deal.id}` : 'deal-customer-name'}
-      >
-        {name}
-      </span>
-      {email ? (
-        <span
-          className="text-xs text-gray-500"
-          data-testid={deal?.id ? `deal-customer-email-${deal.id}` : 'deal-customer-email'}
-        >
-          {email}
-        </span>
-      ) : null}
-      {tags.length ? (
-        <div className="flex flex-wrap gap-1">
-          {tags.map((tag) => (
-            <span
-              key={tag}
-              className="inline-flex items-center rounded-full bg-accent/50 px-2 py-0.5 text-xs text-muted-foreground"
-            >
-              {tag}
-            </span>
-          ))}
-        </div>
-      ) : null}
-    </div>
-  )
-}
-
-const SHEET_CATEGORY_RULES = [
-  { key: 'exterior', tokens: ['exterior', 'paint', 'ppf', 'ceramic', 'coating', 'tint'] },
-  { key: 'interior', tokens: ['interior', 'leather', 'carpet', 'fabric', 'upholstery'] },
-  { key: 'windshield', tokens: ['windshield', 'glass'] },
-  { key: 'rg', tokens: ['rg', 'rust', 'rustguard', 'rust guard', 'rustproof'] },
-]
-
-const normalizeSheetToken = (value) =>
-  String(value || '')
-    .toLowerCase()
-    .trim()
-
-const getSheetDateLabel = (deal) => {
-  const raw =
-    deal?.deal_date ||
-    deal?.dealDate ||
-    deal?.created_at ||
-    deal?.createdAt ||
-    deal?.created ||
-    deal?.inserted_at
-  const date = raw ? toSafeDateForTimeZone(raw) : null
-  if (!date || Number.isNaN(date.getTime())) return '—'
-  return date.toLocaleDateString('en-US', { month: 'numeric', day: 'numeric' })
-}
-
-const getSheetTrackingRef = (deal) => {
-  return (
-    deal?.deal_number ||
-    deal?.dealNumber ||
-    deal?.transaction_number ||
-    deal?.job_number ||
-    deal?.jobNumber ||
-    '—'
-  )
-}
-
-const getSheetSalesLabel = (deal) => {
-  const raw = deal?.sales_consultant_name || deal?.delivery_coordinator_name || ''
-  return raw ? formatStaffName(raw) : '—'
-}
-
-const getSheetCategoryFlags = (deal) => {
-  const parts = Array.isArray(deal?.job_parts) ? deal.job_parts : []
-  const flags = { exterior: false, interior: false, windshield: false, rg: false }
-  const extraLabels = new Map()
-
-  for (const part of parts) {
-    const product = part?.product || part?.products || {}
-    const rawName =
-      product?.name || part?.product_name || part?.productLabel || part?.product_id || ''
-    const rawCategory = product?.category || part?.category || ''
-    const rawOp = product?.op_code || product?.opCode || part?.op_code || ''
-
-    const token = [rawCategory, rawName, rawOp].map(normalizeSheetToken).filter(Boolean).join(' ')
-
-    let matchedKey = null
-    for (const rule of SHEET_CATEGORY_RULES) {
-      if (rule.tokens.some((t) => token.includes(t))) {
-        matchedKey = rule.key
-        flags[rule.key] = true
-        break
-      }
-    }
-
-    if (!matchedKey) {
-      const label = rawOp || abbreviateProductName(rawName)
-      if (label) extraLabels.set(label.toLowerCase(), label)
-    }
-  }
-
-  const extras = Array.from(extraLabels.values())
-  const max = 2
-  const clipped = extras.slice(0, max)
-  const extraCount = Math.max(0, extras.length - clipped.length)
-  const additionalLabel = clipped.length
-    ? `${clipped.join(', ')}${extraCount ? ` +${extraCount}` : ''}`
-    : '—'
-
-  return { flags, additionalLabel }
-}
-
-const SheetSummaryRow = ({ deal, dense = false }) => {
-  const { flags, additionalLabel } = getSheetCategoryFlags(deal)
-  const fin = getDealFinancials(deal)
-  const dateLabel = getSheetDateLabel(deal)
-  const salesLabel = getSheetSalesLabel(deal)
-  const trackingRef = getSheetTrackingRef(deal)
-
-  const flagBadge = (isOn, label) => (
-    <span
-      className={`inline-flex items-center rounded-full border px-2 py-0.5 text-[11px] font-semibold ${
-        isOn
-          ? 'bg-emerald-500/10 text-emerald-200 border-emerald-500/30'
-          : 'bg-card text-gray-500 border-border'
-      }`}
-    >
-      {label}
-    </span>
-  )
-
-  return (
-    <div
-      className={`grid gap-2 text-xs text-muted-foreground ${
-        dense
-          ? 'grid-cols-2 sm:grid-cols-3'
-          : 'grid-cols-12 items-start rounded-lg border border-border bg-card px-3 py-2'
-      }`}
-      data-testid={deal?.id ? `deal-core-snapshot-${deal.id}` : 'deal-core-snapshot'}
-    >
-      <div className={dense ? 'min-w-0' : 'col-span-2'}>
-        <div className="text-[10px] uppercase tracking-wide text-gray-500">Date</div>
-        <div className="font-semibold tabular-nums">{dateLabel}</div>
-      </div>
-      <div className={dense ? 'min-w-0' : 'col-span-3'}>
-        <div className="text-[10px] uppercase tracking-wide text-gray-500">Categories</div>
-        <div className="mt-1 flex flex-wrap gap-1">
-          {flagBadge(flags.exterior, 'Ext')}
-          {flagBadge(flags.interior, 'Int')}
-          {flagBadge(flags.windshield, 'WS')}
-          {flagBadge(flags.rg, 'RG')}
-        </div>
-      </div>
-      <div className={dense ? 'min-w-0' : 'col-span-3'}>
-        <div className="text-[10px] uppercase tracking-wide text-gray-500">Additional</div>
-        <div className="truncate" title={additionalLabel}>
-          {additionalLabel}
-        </div>
-      </div>
-      <div className={dense ? 'min-w-0' : 'col-span-2'}>
-        <div className="text-[10px] uppercase tracking-wide text-gray-500">Sales</div>
-        <div className="truncate" title={salesLabel}>
-          {salesLabel}
-        </div>
-      </div>
-      <div className={dense ? 'min-w-0' : 'col-span-2'}>
-        <div className="text-[10px] uppercase tracking-wide text-gray-500">Tracking</div>
-        <div className="truncate" title={trackingRef}>
-          {trackingRef}
-        </div>
-      </div>
-      <div className={dense ? 'min-w-0' : 'col-span-12'}>
-        <div className="text-[10px] uppercase tracking-wide text-gray-500">
-          Price / Cost / Gross
-        </div>
-        <div className="flex flex-wrap gap-2 font-semibold tabular-nums">
-          <span>S {formatMoney0OrDash(fin.sale)}</span>
-          <span>C {formatMoney0OrDash(fin.cost)}</span>
-          <span>P {formatMoney0OrDash(fin.profit)}</span>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-const SheetViewTable = ({ deals = [], onRowClick }) => {
-  const renderCheck = (value) => (value ? '✓' : '—')
-
-  return (
-    <div className="bg-card rounded-lg border border-border overflow-x-auto">
-      <table className="min-w-[1100px] w-full text-xs">
-        <thead className="bg-card text-muted-foreground uppercase text-[11px] tracking-wide">
-          <tr>
-            <th className="px-3 py-2 text-left font-semibold">Date</th>
-            <th className="px-3 py-2 text-left font-semibold">Customer</th>
-            <th className="px-3 py-2 text-left font-semibold">Vehicle</th>
-            <th className="px-3 py-2 text-center font-semibold">Exterior</th>
-            <th className="px-3 py-2 text-center font-semibold">Interior</th>
-            <th className="px-3 py-2 text-center font-semibold">Windshield</th>
-            <th className="px-3 py-2 text-center font-semibold">RG</th>
-            <th className="px-3 py-2 text-left font-semibold">Additional Package</th>
-            <th className="px-3 py-2 text-right font-semibold">Price</th>
-            <th className="px-3 py-2 text-right font-semibold">Cost</th>
-            <th className="px-3 py-2 text-right font-semibold">Gross</th>
-            <th className="px-3 py-2 text-left font-semibold">Sales</th>
-            <th className="px-3 py-2 text-left font-semibold">Tracking</th>
-          </tr>
-        </thead>
-        <tbody className="divide-y">
-          {deals.map((deal) => {
-            const fin = getDealFinancials(deal)
-            const { flags, additionalLabel } = getSheetCategoryFlags(deal)
-            const vehicle = getDealVehicleDisplay(deal)
-            const customer =
-              deal?.customer_name || deal?.vehicle?.owner_name || deal?.customerEmail || '—'
-
-            return (
-              <tr
-                key={deal?.id}
-                data-testid={`sheet-row-${deal?.id}`}
-                className="hover:bg-accent/10 cursor-pointer"
-                role="button"
-                tabIndex={0}
-                onClick={() => onRowClick?.(deal)}
-                onKeyDown={(e) => {
-                  if (e.key === 'Enter' || e.key === ' ') {
-                    e.preventDefault()
-                    onRowClick?.(deal)
-                  }
-                }}
-              >
-                <td className="px-3 py-2 text-muted-foreground tabular-nums">{getSheetDateLabel(deal)}</td>
-                <td className="px-3 py-2 text-foreground" title={customer}>
-                  {customer}
-                </td>
-                <td className="px-3 py-2 text-muted-foreground" title={vehicle?.title || ''}>
-                  {vehicle?.main || '—'}
-                </td>
-                <td className="px-3 py-2 text-center" data-testid={`sheet-${deal?.id}-exterior`}>
-                  {renderCheck(flags.exterior)}
-                </td>
-                <td className="px-3 py-2 text-center" data-testid={`sheet-${deal?.id}-interior`}>
-                  {renderCheck(flags.interior)}
-                </td>
-                <td className="px-3 py-2 text-center" data-testid={`sheet-${deal?.id}-windshield`}>
-                  {renderCheck(flags.windshield)}
-                </td>
-                <td className="px-3 py-2 text-center" data-testid={`sheet-${deal?.id}-rg`}>
-                  {renderCheck(flags.rg)}
-                </td>
-                <td className="px-3 py-2 text-muted-foreground">{additionalLabel}</td>
-                <td className="px-3 py-2 text-right tabular-nums text-foreground">
-                  {formatMoney0OrDash(fin.sale)}
-                </td>
-                <td className="px-3 py-2 text-right tabular-nums text-foreground">
-                  {formatMoney0OrDash(fin.cost)}
-                </td>
-                <td className="px-3 py-2 text-right tabular-nums text-foreground">
-                  {formatMoney0OrDash(fin.profit)}
-                </td>
-                <td className="px-3 py-2 text-muted-foreground">{getSheetSalesLabel(deal)}</td>
-                <td className="px-3 py-2 text-muted-foreground tabular-nums">
-                  {getSheetTrackingRef(deal)}
-                </td>
-              </tr>
-            )
-          })}
-        </tbody>
-      </table>
-    </div>
-  )
-}
-
-// Determine primary label for a deal card/table row (job number > title > stock > customer)
-
-const getDealNumberDisplay = (deal) => {
-  if (!deal) return '—'
-  return (
-    deal?.job_number ||
-    deal?.jobNumber ||
-    deal?.deal_number ||
-    deal?.dealNumber ||
-    deal?.transaction_number ||
-    '—'
-  )
-}
-
-const getDealDateDisplay = (deal) => {
-  const raw =
-    deal?.deal_date ||
-    deal?.dealDate ||
-    deal?.created_at ||
-    deal?.createdAt ||
-    deal?.created ||
-    deal?.inserted_at
-  return formatCreatedShort(raw) || '—'
-}
-
-const DealCoreSnapshot = ({ deal }) => {
-  if (!deal) return null
-
-  const fin = getDealFinancials(deal)
-  const vehicle = getDealVehicleDisplay(deal)
-  const customer =
-    deal?.customer_name || deal?.customerName || deal?.vehicle?.owner_name || deal?.customer_email || '—'
-  const dealNumber = getDealNumberDisplay(deal)
-  const dateLabel = getDealDateDisplay(deal)
-  const productSummary = getDealProductLabelSummary(deal, 4)
-  const itemsBought =
-    productSummary.labels.length > 0
-      ? `${productSummary.labels.join(', ')}${productSummary.extraCount ? ` +${productSummary.extraCount}` : ''}`
-      : '—'
-
-  return (
-    <div
-      className="rounded-lg border border-border bg-card px-3 py-2"
-      data-testid={deal?.id ? `deal-core-snapshot-${deal.id}` : 'deal-core-snapshot'}
-    >
-      <div className="grid grid-cols-1 gap-2 text-xs sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-6">
-        <div>
-          <div className="text-[10px] uppercase tracking-wide text-gray-500">Customer</div>
-          <div className="truncate font-semibold text-foreground" title={customer}>
-            {customer}
-          </div>
-        </div>
-        <div>
-          <div className="text-[10px] uppercase tracking-wide text-gray-500">Vehicle</div>
-          <div className="truncate font-semibold text-foreground" title={vehicle?.title || ''}>
-            {vehicle?.isMissing ? '—' : vehicle?.main}
-          </div>
-        </div>
-        <div>
-          <div className="text-[10px] uppercase tracking-wide text-gray-500">Date</div>
-          <div className="font-semibold tabular-nums text-foreground">{dateLabel}</div>
-        </div>
-        <div>
-          <div className="text-[10px] uppercase tracking-wide text-gray-500">Deal #</div>
-          <div className="truncate font-semibold tabular-nums text-foreground" title={dealNumber}>
-            {dealNumber}
-          </div>
-        </div>
-        <div>
-          <div className="text-[10px] uppercase tracking-wide text-gray-500">Profit</div>
-          <div className="font-semibold tabular-nums text-foreground">
-            {formatMoney0OrDash(fin?.profit)}
-          </div>
-        </div>
-        <div>
-          <div className="text-[10px] uppercase tracking-wide text-gray-500">Items Bought</div>
-          <div className="truncate font-semibold text-foreground" title={itemsBought}>
-            {itemsBought}
-          </div>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-const getDealPrimaryRef = (deal) => {
-  if (!deal) return 'Deal'
-
-  const jobNumber = deal?.job_number || deal?.jobNumber
-  const title = (deal?.title || deal?.description || '').trim()
-  const stockNumber =
-    deal?.vehicle?.stock_number || deal?.stock_no || deal?.stockNumber || deal?.vehicle_stock
-  const fallbackId = deal?.id ? `Job-${String(deal.id).slice(0, 8)}` : ''
-  const customer = (deal?.customer_name || deal?.customerName || '').trim()
-
-  if (jobNumber && title) return `${jobNumber} • ${title}`
-  if (jobNumber && stockNumber) return `${jobNumber} • Stock ${stockNumber}`
-  if (jobNumber) return jobNumber
-  if (title && stockNumber) return `${title} • Stock ${stockNumber}`
-  if (title) return title
-  if (stockNumber) return `Stock ${stockNumber}`
-  if (customer) return customer
-  return fallbackId || 'Deal'
-}
-
-const toFiniteNumberOrNull = (value) => {
-  if (value == null) return null
-  const num = typeof value === 'number' ? value : Number.parseFloat(value)
-  return Number.isFinite(num) ? num : null
-}
-
-const formatMoney0OrDash = (value) => {
-  const num = toFiniteNumberOrNull(value)
-  return num == null ? '—' : money0.format(num)
-}
-
-// NOTE: `getDealFinancials` is imported from `src/utils/dealKpis.js` to keep
-// Deals + Analytics KPI computation consistent.
-
-// ✅ UPDATED: Service Location Tag with color styling per requirements
-const ServiceLocationTag = ({ jobParts }) => {
-  // Check if any line items are off-site to determine vendor status
-  const hasOffSiteItems = jobParts?.some((part) => part?.is_off_site)
-  const hasOnSiteItems = jobParts?.some((part) => !part?.is_off_site)
-
-  if (hasOffSiteItems && hasOnSiteItems) {
-    return (
-      <div className="flex flex-col space-y-1">
-        <span className="inline-flex items-center px-2 py-1 rounded text-xs font-medium bg-accent/50 text-foreground border border-border">
-          Off-Site
-        </span>
-        <span className="inline-flex items-center px-2 py-1 rounded text-xs font-medium bg-accent/50 text-foreground border border-border">
-          On-Site
-        </span>
-      </div>
-    )
-  }
-
-  if (hasOffSiteItems) {
-    return (
-      <span className="inline-flex items-center px-2 py-1 rounded text-xs font-medium bg-accent/50 text-foreground border border-border">
-        Off-Site
-      </span>
-    )
-  }
-
-  return (
-    <span className="inline-flex items-center px-2 py-1 rounded text-xs font-medium bg-accent/50 text-foreground border border-border">
-      On-Site
-    </span>
-  )
-}
-
-// ✅ UPDATED: Enhanced draft reminder with improved styling
-const DraftReminderBanner = ({ draftsCount, onViewDrafts }) => {
-  const [dismissed, setDismissed] = useState(false)
-
-  if (draftsCount === 0 || dismissed) return null
-
-  return (
-    <div
-      className="mb-6 p-4 rounded-lg border"
-      style={{ backgroundColor: 'rgba(245, 158, 11, 0.12)', borderColor: 'rgba(245, 158, 11, 0.35)', color: '#FCD34D' }}
-    >
-      <div className="flex items-center justify-between">
-        <div className="flex items-center space-x-3">
-          <div className="flex-shrink-0">
-            <Icon name="AlertCircle" size={20} style={{ color: '#F59E0B' }} />
-          </div>
-          <div>
-            <p className="font-medium">Draft – needs details</p>
-            <p className="text-sm">
-              You have {draftsCount} draft deal{draftsCount > 1 ? 's' : ''} to complete.
-            </p>
-          </div>
-        </div>
-        <div className="flex items-center space-x-2">
-          <Button
-            size="sm"
-            variant="ghost"
-            onClick={onViewDrafts}
-            style={{ color: '#FCD34D' }}
-            className="hover:bg-yellow-500/10"
-            aria-label="View draft deals"
-          >
-            View drafts
-          </Button>
-          <button onClick={() => setDismissed(true)} className="p-1" style={{ color: '#FCD34D' }}>
-            <Icon name="X" size={16} />
-          </button>
-        </div>
-      </div>
-    </div>
-  )
-}
 
 export default function DealsPage() {
   const toast = useToast?.()
@@ -1267,34 +552,6 @@ export default function DealsPage() {
     setShowDetailDrawer(true)
   }
 
-  // ✅ FIXED: Enhanced error display component
-  const ErrorAlert = ({ message, onClose }) => {
-    if (!message) return null
-
-    return (
-      <div className="mb-4 p-4 bg-red-50 border border-red-200 rounded-lg">
-        <div className="flex justify-between items-start">
-          <div className="flex">
-            <Icon name="AlertCircle" size={20} className="text-red-500 mr-2 mt-0.5" />
-            <div>
-              <h4 className="text-sm font-medium text-red-800">Error</h4>
-              <p className="text-sm text-red-700 mt-1">{message}</p>
-            </div>
-          </div>
-          {onClose && (
-            <button
-              onClick={onClose}
-              className="text-red-400 hover:text-red-600"
-              aria-label="Dismiss error"
-            >
-              <Icon name="X" size={16} />
-            </button>
-          )}
-        </div>
-      </div>
-    )
-  }
-
   // ✅ UPDATED: Enhanced filter deals with 300ms debounced search
   const filteredDeals = deals?.filter((deal) => {
     // Status filter with tab-based logic
@@ -1721,450 +978,22 @@ export default function DealsPage() {
           onViewDrafts={() => updateFilter('status', 'Draft')}
         />
 
-        {/* ✅ UPDATED: KPI Row - Enhanced with profit analysis */}
-        <div className="mb-6" data-testid="kpi-row">
-          <div className="grid grid-cols-2 md:grid-cols-6 gap-4">
-            {/* Active Jobs */}
-            <div className="bg-card p-6 rounded-xl border border-border shadow-sm">
-              <div className="flex items-center">
-                <div className="p-3 rounded-lg bg-accent/50 mr-4">
-                  <Icon name="Clock" size={24} className="text-foreground" />
-                </div>
-                <div>
-                  <h3 className="text-muted-foreground text-sm font-medium uppercase tracking-wide">
-                    Active
-                  </h3>
-                  <p className="text-foreground text-2xl font-bold">{kpis?.active}</p>
-                </div>
-              </div>
-            </div>
+        {/* ✅ UPDATED: KPI Row */}
+        <DealsKpiRow kpis={kpis} />
 
-            {/* Revenue */}
-            <div className="bg-card p-6 rounded-xl border border-border shadow-sm">
-              <div className="flex items-center">
-                <div className="p-3 rounded-lg bg-accent/50 mr-4">
-                  <Icon name="DollarSign" size={24} className="text-foreground" />
-                </div>
-                <div>
-                  <h3 className="text-muted-foreground text-sm font-medium uppercase tracking-wide">
-                    Revenue
-                  </h3>
-                  <p className="text-foreground text-2xl font-bold">
-                    {money0.format(parseFloat(kpis?.revenue) || 0)}
-                  </p>
-                </div>
-              </div>
-            </div>
-
-            {/* Profit */}
-            <div className="bg-card p-6 rounded-xl border border-border shadow-sm">
-              <div className="flex items-center">
-                <div className="p-3 rounded-lg bg-accent/50 mr-4">
-                  <Icon name="TrendingUp" size={24} className="text-foreground" />
-                </div>
-                <div>
-                  <h3 className="text-muted-foreground text-sm font-medium uppercase tracking-wide">
-                    Profit
-                  </h3>
-                  <p className="text-foreground text-2xl font-bold">
-                    {kpis?.profit === '' || kpis?.profit == null
-                      ? '—'
-                      : money0.format(parseFloat(kpis?.profit) || 0)}
-                  </p>
-                </div>
-              </div>
-            </div>
-
-            {/* Margin */}
-            <div className="bg-card p-6 rounded-xl border border-border shadow-sm">
-              <div className="flex items-center">
-                <div className="p-3 rounded-lg bg-accent/50 mr-4">
-                  <Icon name="Percent" size={24} className="text-foreground" />
-                </div>
-                <div>
-                  <h3 className="text-muted-foreground text-sm font-medium uppercase tracking-wide">
-                    Margin
-                  </h3>
-                  <p className="text-foreground text-2xl font-bold">
-                    {kpis?.margin === '' || kpis?.margin == null
-                      ? '—'
-                      : pct1(parseFloat(kpis?.margin) / 100 || 0)}
-                  </p>
-                </div>
-              </div>
-            </div>
-
-            {/* Needs Work (time TBD) */}
-            <div className="bg-card p-6 rounded-xl border border-border shadow-sm">
-              <div className="flex items-center">
-                <div className="p-3 rounded-lg bg-accent/50 mr-4">
-                  <Icon name="Clock" size={24} className="text-foreground" />
-                </div>
-                <div>
-                  <h3 className="text-muted-foreground text-sm font-medium uppercase tracking-wide">
-                    Needs Work (time TBD)
-                  </h3>
-                  <p className="text-foreground text-2xl font-bold">{kpis?.pending}</p>
-                </div>
-              </div>
-            </div>
-
-            {/* Products Sold */}
-            <div className="bg-card p-6 rounded-xl border border-border shadow-sm">
-              <div className="flex items-center">
-                <div className="p-3 rounded-lg bg-accent/50 mr-4">
-                  <Icon name="Package" size={24} className="text-foreground" />
-                </div>
-                <div>
-                  <h3 className="text-muted-foreground text-sm font-medium uppercase tracking-wide">
-                    Items Sold
-                  </h3>
-                  <p className="text-foreground text-2xl font-bold">{kpis?.productsSold || 0}</p>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-
-        {/* Status tabs and search (advanced dropdowns removed) */}
-        <div className="mb-6 bg-card rounded-lg border border-border p-4">
-          {/* Status Tabs */}
-          <div className="flex flex-wrap gap-2 mb-4">
-            {[
-              { value: 'All', label: 'All' },
-              { value: 'Open', label: 'Open' },
-              { value: 'Draft', label: 'Draft' },
-              { value: 'Pending', label: 'Needs Work (time TBD)' },
-              { value: 'Scheduled', label: 'Scheduled' },
-              { value: 'Active', label: 'Active' },
-              { value: 'Completed', label: 'Completed' },
-            ]?.map(({ value, label }) => (
-              <button
-                key={value}
-                onClick={() => updateFilter('status', value)}
-                className={`px-4 py-2 rounded-lg text-sm font-medium transition-colors
-                  ${
-                    filters?.status === value
-                      ? '0 text-white border border-border'
-                      : 'bg-transparent border border-border text-muted-foreground hover:bg-card'
-                  }`}
-              >
-                {label}
-              </button>
-            ))}
-          </div>
-
-          {/* Preset Views (hidden per request) */}
-          {false && (
-            <div className="flex flex-wrap gap-2 mb-4">
-              {[
-                'All',
-                'Today',
-                'Past Due',
-                'Unscheduled',
-                'Off-site Today',
-                'Awaiting Vendor/Parts',
-                'Completed—awaiting pickup',
-                'My Deals',
-                'Loaners Out',
-                'Loaners Due',
-                'Loaners Overdue',
-              ]?.map((view) => (
-                <button
-                  key={view}
-                  onClick={() => updateFilter('presetView', view)}
-                  className={`px-3 py-1.5 rounded-lg text-xs font-medium transition-colors border
-                    ${
-                      filters?.presetView === view
-                        ? '0 text-white border-border'
-                        : 'bg-transparent text-muted-foreground border-border hover:bg-card'
-                    }`}
-                >
-                  {view}
-                </button>
-              ))}
-            </div>
-          )}
-
-          <div className="flex flex-col lg:flex-row gap-4">
-            {/* ✅ UPDATED: Search box with 300ms debounce, matches stock, name, phone (strip non-digits) */}
-            <div className="flex-1">
-              <div className="relative">
-                <Icon
-                  name="Search"
-                  size={16}
-                  className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-500"
-                />
-                <input
-                  type="text"
-                  placeholder="Search deals, customers, vehicles..."
-                  value={filters?.search}
-                  onChange={(e) => updateFilter('search', e?.target?.value)}
-                  className="bg-background border border-border rounded-lg w-full h-11 pl-9 pr-3 text-foreground placeholder:text-gray-500 focus:ring-2 focus:ring-[rgb(var(--ring)/0.35)] focus:border-transparent"
-                />
-              </div>
-            </div>
-
-            <div className="flex flex-wrap items-center gap-2">
-              <label className="text-xs font-medium text-muted-foreground" htmlFor="deals-month-filter">
-                Month
-              </label>
-              <input
-                id="deals-month-filter"
-                type="month"
-                value={filters?.createdMonth || ''}
-                onChange={(e) => updateFilter('createdMonth', e?.target?.value || '')}
-                className="h-11 rounded-lg border border-border bg-background px-3 text-sm text-foreground"
-              />
-              <Button
-                type="button"
-                variant="ghost"
-                size="sm"
-                onClick={() => updateFilter('createdMonth', '')}
-                className="text-muted-foreground hover:text-foreground"
-              >
-                All months
-              </Button>
-            </div>
-
-            <details className="group relative">
-              <summary className="flex cursor-pointer list-none items-center gap-2 rounded-lg border border-border bg-card px-3 py-2 text-sm text-muted-foreground">
-                <Icon name="Filter" size={14} /> Filters
-              </summary>
-              <div className="absolute right-0 z-30 mt-2 w-64 rounded-lg border border-border bg-background p-3 text-xs text-muted-foreground shadow-lg">
-                <label
-                  className="block text-[11px] font-semibold text-muted-foreground"
-                  htmlFor="deals-location-filter"
-                >
-                  Location
-                </label>
-                <select
-                  id="deals-location-filter"
-                  className="mt-2 h-9 w-full rounded-md border border-border bg-background px-2 text-sm text-foreground"
-                  value={filters?.location || 'All'}
-                  onChange={(e) => updateFilter('location', e?.target?.value || 'All')}
-                >
-                  <option value="All">All</option>
-                  <option value="In-House">In-House</option>
-                  <option value="Off-Site">Off-Site</option>
-                  <option value="Mixed">Mixed</option>
-                </select>
-
-                <label
-                  className="mt-3 block text-[11px] font-semibold text-muted-foreground"
-                  htmlFor="deals-loaner-filter"
-                >
-                  Loaner
-                </label>
-                <select
-                  id="deals-loaner-filter"
-                  className="mt-2 h-9 w-full rounded-md border border-border bg-background px-2 text-sm text-foreground"
-                  value={filters?.loanerStatus || 'All'}
-                  onChange={(e) => updateFilter('loanerStatus', e?.target?.value || 'All')}
-                >
-                  <option value="All">All</option>
-                  <option value="Active">Active</option>
-                  <option value="Due Today">Due Today</option>
-                  <option value="Overdue">Overdue</option>
-                  <option value="None">None</option>
-                </select>
-              </div>
-            </details>
-
-            {/* Clear Filters */}
-            <div className="flex items-center">
-              <Button
-                variant="ghost"
-                size="sm"
-                onClick={clearAllFilters}
-                className="text-muted-foreground hover:text-foreground"
-                aria-label="Clear all filters"
-              >
-                <Icon name="X" size={16} className="mr-1" />
-                Clear
-              </Button>
-            </div>
-
-            {/* View Toggle (desktop) */}
-            <div className="hidden md:flex items-center gap-2">
-              <button
-                type="button"
-                onClick={() => setShowSheetView(false)}
-                className={`h-9 px-3 rounded-md text-xs font-medium border transition-colors ${
-                  showSheetView
-                    ? 'bg-card text-muted-foreground border-border 0'
-                    : '0 text-white border-border'
-                }`}
-                aria-pressed={!showSheetView}
-              >
-                Card View
-              </button>
-              <button
-                type="button"
-                onClick={() => setShowSheetView(true)}
-                className={`h-9 px-3 rounded-md text-xs font-medium border transition-colors ${
-                  showSheetView
-                    ? '0 text-white border-border'
-                    : 'bg-card text-muted-foreground border-border 0'
-                }`}
-                aria-pressed={showSheetView}
-              >
-                Sheet View
-              </button>
-            </div>
-          </div>
-
-          {/* Advanced filter dropdowns removed; search covers all filtering needs */}
-          {false && (
-            <div className="mt-4 grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
-              {/* Vendor */}
-              <div>
-                <label className="block text-xs font-medium text-muted-foreground mb-1">Vendor</label>
-                <select
-                  className="bg-background border border-border rounded-lg w-full h-11 px-3 text-foreground"
-                  value={filters.vendor || ''}
-                  onChange={(e) => updateFilter('vendor', e.target.value || null)}
-                >
-                  <option value="">All</option>
-                  {vendorOptions.map((v) => (
-                    <option key={v.value} value={v.value}>
-                      {v.label}
-                    </option>
-                  ))}
-                </select>
-              </div>
-
-              {/* Sales */}
-              <div>
-                <label className="block text-xs font-medium text-muted-foreground mb-1">Sales</label>
-                <select
-                  className="bg-background border border-border rounded-lg w-full h-11 px-3 text-foreground"
-                  value={filters.salesAssigned || ''}
-                  onChange={(e) => updateFilter('salesAssigned', e.target.value || null)}
-                >
-                  <option value="">All</option>
-                  {salesOptions.map((o) => (
-                    <option key={o.value} value={o.value}>
-                      {o.label}
-                    </option>
-                  ))}
-                </select>
-              </div>
-
-              {/* Finance */}
-              <div>
-                <label className="block text-xs font-medium text-muted-foreground mb-1">Finance</label>
-                <select
-                  className="bg-background border border-border rounded-lg w-full h-11 px-3 text-foreground"
-                  value={filters.financeAssigned || ''}
-                  onChange={(e) => updateFilter('financeAssigned', e.target.value || null)}
-                >
-                  <option value="">All</option>
-                  {financeOptions.map((o) => (
-                    <option key={o.value} value={o.value}>
-                      {o.label}
-                    </option>
-                  ))}
-                </select>
-              </div>
-
-              {/* Delivery */}
-              <div>
-                <label className="block text-xs font-medium text-muted-foreground mb-1">Delivery</label>
-                <select
-                  className="bg-background border border-border rounded-lg w-full h-11 px-3 text-foreground"
-                  value={filters.deliveryAssigned || ''}
-                  onChange={(e) => updateFilter('deliveryAssigned', e.target.value || null)}
-                >
-                  <option value="">All</option>
-                  {deliveryOptions.map((o) => (
-                    <option key={o.value} value={o.value}>
-                      {o.label}
-                    </option>
-                  ))}
-                </select>
-              </div>
-
-              {/* Location */}
-              <div>
-                <label className="block text-xs font-medium text-muted-foreground mb-1">Location</label>
-                <select
-                  className="bg-background border border-border rounded-lg w-full h-11 px-3 text-foreground"
-                  value={filters.location}
-                  onChange={(e) => updateFilter('location', e.target.value)}
-                >
-                  {['All', 'In-House', 'Off-Site', 'Mixed'].map((opt) => (
-                    <option key={opt} value={opt}>
-                      {opt}
-                    </option>
-                  ))}
-                </select>
-              </div>
-
-              {/* Loaner */}
-              <div>
-                <label className="block text-xs font-medium text-muted-foreground mb-1">Loaner</label>
-                <select
-                  className="bg-background border border-border rounded-lg w-full h-11 px-3 text-foreground"
-                  value={filters.loanerStatus}
-                  onChange={(e) => updateFilter('loanerStatus', e.target.value)}
-                >
-                  {['All', 'Active', 'Due Today', 'Overdue', 'None'].map((opt) => (
-                    <option key={opt} value={opt}>
-                      {opt}
-                    </option>
-                  ))}
-                </select>
-              </div>
-
-              {/* Promise Date Range */}
-              <div className="grid grid-cols-2 gap-2">
-                <div>
-                  <label className="block text-xs font-medium text-muted-foreground mb-1">
-                    Promise from
-                  </label>
-                  <input
-                    type="date"
-                    className="bg-background border border-border rounded-lg w-full h-11 px-3 text-foreground"
-                    value={filters.promiseStartDate}
-                    onChange={(e) => updateFilter('promiseStartDate', e.target.value)}
-                  />
-                </div>
-                <div>
-                  <label className="block text-xs font-medium text-muted-foreground mb-1">
-                    Promise to
-                  </label>
-                  <input
-                    type="date"
-                    className="bg-background border border-border rounded-lg w-full h-11 px-3 text-foreground"
-                    value={filters.promiseEndDate}
-                    onChange={(e) => updateFilter('promiseEndDate', e.target.value)}
-                  />
-                </div>
-              </div>
-
-              {/* Work tags (multi-select) */}
-              <div>
-                <label className="block text-xs font-medium text-muted-foreground mb-1">Work tags</label>
-                <select
-                  multiple
-                  className="bg-background border border-border rounded-lg w-full min-h-[44px] px-3 py-2 text-foreground"
-                  value={filters.workTags}
-                  onChange={(e) => {
-                    const selected = Array.from(e.target.selectedOptions).map((o) => o.value)
-                    updateFilter('workTags', selected)
-                  }}
-                >
-                  {allWorkTags.map((t) => (
-                    <option key={t} value={t}>
-                      {t}
-                    </option>
-                  ))}
-                </select>
-              </div>
-            </div>
-          )}
-        </div>
+        {/* Status tabs and search */}
+        <DealsFilterBar
+          filters={filters}
+          updateFilter={updateFilter}
+          clearAllFilters={clearAllFilters}
+          showSheetView={showSheetView}
+          setShowSheetView={setShowSheetView}
+          vendorOptions={vendorOptions}
+          salesOptions={salesOptions}
+          deliveryOptions={deliveryOptions}
+          financeOptions={financeOptions}
+          allWorkTags={allWorkTags}
+        />
 
         {/* Results count */}
         <div className="mb-4 text-sm text-muted-foreground flex items-center gap-3">
@@ -2830,42 +1659,13 @@ export default function DealsPage() {
         />
 
         {/* ✅ UPDATED: Delete Confirmation Modal */}
-        {deleteConfirm && (
-          <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50 p-4">
-            <div className="bg-background border border-border rounded-lg w-full max-w-md max-h-[80vh] overflow-y-auto p-4">
-              <div className="p-6">
-                <h3 className="text-lg font-semibold mb-4 text-foreground">Delete Deal</h3>
-                <p className="text-muted-foreground mb-6">
-                  Delete deal and its line items? This cannot be undone.
-                </p>
-                {!!error && (
-                  <div className="mb-4 p-3 bg-red-500/10 border border-red-500/20 rounded">
-                    <div className="text-sm text-red-200">{error}</div>
-                  </div>
-                )}
-                <div className="flex gap-3">
-                  <Button
-                    variant="outline"
-                    onClick={() => setDeleteConfirm(null)}
-                    className="flex-1 h-11 bg-card border border-border text-foreground 0"
-                    aria-label="Cancel deletion"
-                    disabled={deletingDeal}
-                  >
-                    Cancel
-                  </Button>
-                  <Button
-                    onClick={() => handleDeleteDeal(deleteConfirm?.id)}
-                    className="flex-1 h-11 bg-red-600 hover:bg-red-700 text-white"
-                    aria-label="Confirm deletion"
-                    disabled={deletingDeal}
-                  >
-                    {deletingDeal ? 'Deleting...' : 'Delete'}
-                  </Button>
-                </div>
-              </div>
-            </div>
-          </div>
-        )}
+        <DeleteConfirmModal
+          deleteConfirm={deleteConfirm}
+          error={error}
+          deletingDeal={deletingDeal}
+          onDelete={handleDeleteDeal}
+          onCancel={() => setDeleteConfirm(null)}
+        />
 
         {/* Deal Detail Drawer (read-only) */}
         <DealDetailDrawer
